+++ app/controllers/api/v1/users/current/twitter_data_controller.rb
@@ -2,7 +2,7 @@ class Api::V1::Users::Current::TwitterDataController < Api::V1::BaseController
   before_action :require_login
 
   def update
-    data = TwitterData::User.new(current_user).call
+    data = TwitterAPI::User.new(current_user).call
     current_user.update!(name: data[:name],
                          screen_name: data[:screen_name],
                          avatar_url: data[:avatar_url])
